[BACKPORT 2025.1][PLAT-16470] Use long to store query call and row count
Summary:
**Context**
We've recently hit a case where the total query call count exceeded the range of a Java int.
This interger overflow caused the call stat (and related derived stats) to show up as negative numbers.

**Change**
Since we may overflow a Java int, this diff updates the data type to a long.

Original commit: 92e26915a88d6cc7d9778deb615f5886435cd1ff / D44246

Test Plan:
As a sanity check, run some heavy workflow and ensure the slow query stats appear as expected.
Then verify that we don't overflow when the total call count or row count exceed the range of a Java int.
{F359070}

Reviewers: asharma, amalyshev, skurapati

Reviewed By: amalyshev

Differential Revision: https://phorge.dev.yugabyte.com/D45060

+++ managed/src/main/java/com/yugabyte/yw/queries/QueryHelper.java
@@ -339,16 +339,17 @@ public class QueryHelper {
                   // Defining values to reuse
                   double X_a = previousQueryObj.get("mean_time").asDouble();
                   double X_b = queryObject.get("mean_time").asDouble();
-                  int n_a = previousQueryObj.get("calls").asInt();
-                  int n_b = queryObject.get("calls").asInt();
+                  long n_a = previousQueryObj.get("calls").asLong();
+                  long n_b = queryObject.get("calls").asLong();
                   double S_a = previousQueryObj.get("stddev_time").asDouble();
                   double S_b = queryObject.get("stddev_time").asDouble();
 
                   double totalTime =
                       previousQueryObj.get("total_time").asDouble()
                           + queryObject.get("total_time").asDouble();
-                  int totalCalls = n_a + n_b;
-                  int rows = previousQueryObj.get("rows").asInt() + queryObject.get("rows").asInt();
+                  long totalCalls = n_a + n_b;
+                  long rows =
+                      previousQueryObj.get("rows").asLong() + queryObject.get("rows").asLong();
                   double minTime =
                       Math.min(
                           previousQueryObj.get("min_time").asDouble(),
